// src/app/models/coach/coach.ts
export class Coach {
    id: string;
    create: Date;  
    userCreate: string;
    firstname: string;
    lastname: string;
    birthday: Date;
    username: string;
    password: string;
    address: string;
    email: string;    
    phone: string;
    movil:string;
    club: string;
    estatus:boolean;
}